requirements.txt
pyserial~=3.5
PyAudio~=0.2.13
numpy~=1.24.2
keyboard~=0.13.5
